mysite/requirements.txt
Django==1.6.5
Markdown==2.4.1
django-filter==0.7
djangorestframework==2.4.2
